nixos/oci-containers: let systemd generate a podman cidfile dynamically
Hardcoding the cidfile to /run/ does not work for containers running as a specific user without root previleges. When podman compose generates a systemd service, it instead uses /%t/%n-cid as the file path, which lets systemd dynamically
select a path according to the rules defined in the SPECIFIERS section of it's man page. Unfortunately, this is not possible when using a script file since systemd won't resolve wildcards in it, thus we have to move the execution into the
systemd unit.