nixos/oci-containers: refactor podman config to generate cidfile in systemd unit
When run as a user other than root, podman is unable to write it's container-id-file into /run/podman because it has no permissions to it. The strategy employed by podman-compose is to use systemd specifiers to provide a writeable directory depending on context. The specifiers used are %t and %n. See https://www.freedesktop.org/software/systemd/man/systemd.unit.html#Specifiers for documentation.